requirements.txt
anyjson==0.3.3
boto==2.42.0
defusedxml==0.4.1
dj-database-url==0.2.2
dj-static==0.0.6
Django==1.10
django-bootstrap3==7.0.1
django-filter==0.15.2
django-fontawesome==0.2.6
django-storages==1.5.0
djangorestframework==3.3.3
gunicorn==18.0
mock==2.0.0
oauthlib==1.1.2
pbr==1.10.0
Pillow==3.3.1
psycopg2==2.6.2
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
PyYAML==3.11
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
static3==0.5.1
travis==0.0.2
djangorestframework==3.3.3
anyjson==0.3.3
Pillow==3.3.1
django-rest-framework-social-oauth2==1.0.4
